move download.js, stringify.js into scripts/ folder and fix imports
